+++ NEWS.md
@@ -8,8 +8,13 @@ subtitle: plm - Linear Models for Panel Data - A set of estimators and tests for
 
 # plm 2.6-9999
 
-* Changes relative to CRAN-released version 2.6-2:
-(none so far)
+Changes relative to CRAN-released version 2.6-2:
+
+### Speed-up:
+* Replaced base R's `duplicated` with faster `collapse::fduplicated(x, all = FALSE)`.
+
+***
+
 
 # plm 2.6-3